+++ src/nodes/Nodes/Shapes/Definitions/Box2dPolygonDefNode.cpp
@@ -60,10 +60,10 @@ namespace VVVV
 							vertidx = 0;
 						}
 					}
-					shapeDef->density = density;
-					shapeDef->friction = friction;
-					shapeDef->restitution = restitution;
-					shapeDef->isSensor = issensor >= 0.5;
+					shapeDef->density = (float)density;
+					shapeDef->friction = (float)friction;
+					shapeDef->restitution = (float)restitution;
+					shapeDef->isSensor = (float)issensor >= 0.5;
 					shapeDef->filter.groupIndex = System::Convert::ToInt32(group);
 					this->m_shapes->AddCustom(custom);
 				}
